+++ web/src/Routes.js
@@ -2,17 +2,16 @@ import React from 'react'
 import { Route, Switch, BrowserRouter as Router } from 'react-router-dom'
 import Test from './screens/test'
 import RoleShow from './screens/roleShow'
-const Sample = ()=>{
-  return <h1>Sample Functional Component</h1>
-}
+
+const Sample = () => (<h1>Sample Functional Component</h1>)
 
 export default () => (
   <Router>
     <Switch>
-      <Route exact path="/" exact component={ Test } />
-      <Route exact path="/front-end" exact component={ RoleShow } />
-      <Route exact path="/back-end" exact component={ Sample } />
-      <Route exact path="/qa" exact component={ Sample } />
+      <Route exact path="/" component={ Test } />
+      <Route exact path="/front-end" component={ RoleShow } />
+      <Route exact path="/back-end" component={ Sample } />
+      <Route exact path="/qa" component={ Sample } />
     </Switch>
   </Router>
 )
