+++ web/src/constants/JobInfo.js
@@ -1,0 +1,34 @@
+const JobInfo = {
+    title: "Junior React Developer",
+    data: {
+        topSkills: ["Job Target: 1", "Job Target: 2", "Job Target: 3","Job Target: 4"],
+        prefSkills: ["Soft Skill: 1", "Soft Skill: 2", "Soft Skill: 3","Soft Skill: 4"]}
+        ,
+    label: { 
+            category: 'Front End', 
+            role: "Engineer"
+        },
+    imgArr : [
+        {   
+            alt: "Slide 1",
+            src: "https://www.iconsdb.com/icons/preview/white/pie-chart-xxl.png",
+            term: "Last Year"
+        },
+        {   alt: "Slide 2",
+            src: "https://www.iconsdb.com/icons/preview/white/pie-chart-xxl.png",
+            term: "Last Quarter"
+        },
+        {   
+            alt: "Slide 3",
+            src: "https://www.iconsdb.com/icons/preview/white/pie-chart-xxl.png",
+            term: "Last Month"
+        }
+    ],
+    style : {
+        header : "card-subtitle center my-2 md-6",
+        graphics: "",
+        summary: "card-title center md-6 my-3"
+    }
+}
+
+export default JobInfo
