+++ web/src/constants/index.js
@@ -1,0 +1,4 @@
+import JobInfo from './JobInfo'
+
+
+export {JobInfo}
